Feature/board : Add Ship placement and grid display
